+++ backend/src/main/java/com/example/stamp_app/controller/MicroControllerController.java
@@ -1,10 +1,9 @@
 package com.example.stamp_app.controller;
 
-import com.example.stamp_app.controller.Param.MicroControllerPostParam;
-import com.example.stamp_app.controller.Response.MicroControllerPostResponse;
+import com.example.stamp_app.controller.param.MicroControllerPostParam;
+import com.example.stamp_app.controller.response.MicroControllerPostResponse;
 import com.example.stamp_app.service.MicroControllerService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
