--- backend/src/main/java/com/example/stamp_app/controller/Response/AccountGetResponse.java
@@ -1,19 +1,0 @@
-package com.example.stamp_app.controller.Response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import org.springframework.http.HttpStatus;
-
-import java.math.BigInteger;
-
-@Data
-@AllArgsConstructor
-public class AccountGetResponse {
-
-    private HttpStatus status;
-
-    private BigInteger id;
-
-    private String name;
-
-}
