+++ backend/src/main/java/com/example/stamp_app/controller/param/MicroControllerPostParam.java
@@ -1,4 +1,4 @@
-package com.example.stamp_app.controller.Param;
+package com.example.stamp_app.controller.param;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
