+++ backend/src/main/java/com/example/stamp_app/controller/response/AccountGetResponse.java
@@ -1,6 +1,5 @@
-package com.example.stamp_app.controller.Response;
+package com.example.stamp_app.controller.response;
 
-import com.example.stamp_app.entity.Account;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.springframework.http.HttpStatus;
@@ -9,10 +8,12 @@ import java.math.BigInteger;
 
 @Data
 @AllArgsConstructor
-public class AccountLoginResponse {
+public class AccountGetResponse {
 
     private HttpStatus status;
 
-    private Account account;
+    private BigInteger id;
+
+    private String name;
 
 }
