+++ backend/src/main/java/com/example/stamp_app/controller/response/AccountLoginResponse.java
@@ -1,0 +1,16 @@
+package com.example.stamp_app.controller.response;
+
+import com.example.stamp_app.dummyData.Account;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import org.springframework.http.HttpStatus;
+
+@Data
+@AllArgsConstructor
+public class AccountLoginResponse {
+
+    private HttpStatus status;
+
+    private Account account;
+
+}
