+++ backend/src/main/java/com/example/stamp_app/controller/response/MicroControllerPostResponse.java
@@ -1,6 +1,6 @@
-package com.example.stamp_app.controller.Response;
+package com.example.stamp_app.controller.response;
 
-import com.example.stamp_app.entity.MicroController;
+import com.example.stamp_app.dummyData.MicroController;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.springframework.http.HttpStatus;
