+++ backend/src/main/java/com/example/stamp_app/dummyData/EnvironmentalData.java
@@ -1,6 +1,6 @@
-package com.example.stamp_app.entity;
+package com.example.stamp_app.dummyData;
 
-import com.example.stamp_app.controller.Param.EnvironmentalDataParam;
+import com.example.stamp_app.controller.param.EnvironmentalDataParam;
 import lombok.Data;
 
 import jakarta.persistence.*;
