+++ backend/src/main/java/com/example/stamp_app/dummyData/MeasuredDataMaster.java
@@ -1,4 +1,4 @@
-package com.example.stamp_app.entity;
+package com.example.stamp_app.dummyData;
 
 import lombok.Data;
