+++ backend/src/main/java/com/example/stamp_app/repository/AccountRepository.java
@@ -1,6 +1,6 @@
 package com.example.stamp_app.repository;
 
-import com.example.stamp_app.entity.Account;
+import com.example.stamp_app.dummyData.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
