+++ backend/src/main/java/com/example/stamp_app/repository/EnvironmentalDataRepository.java
@@ -1,6 +1,6 @@
 package com.example.stamp_app.repository;
 
-import com.example.stamp_app.entity.EnvironmentalData;
+import com.example.stamp_app.dummyData.EnvironmentalData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
