+++ backend/src/main/java/com/example/stamp_app/repository/MeasuredDataMasterRepository.java
@@ -1,6 +1,6 @@
 package com.example.stamp_app.repository;
 
-import com.example.stamp_app.entity.MeasuredDataMaster;
+import com.example.stamp_app.dummyData.MeasuredDataMaster;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
