+++ backend/src/main/java/com/example/stamp_app/repository/Sdi12DataRepository.java
@@ -1,6 +1,6 @@
 package com.example.stamp_app.repository;
 
-import com.example.stamp_app.entity.Sdi12Data;
+import com.example.stamp_app.dummyData.Sdi12Data;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
