+++ backend/src/main/java/com/example/stamp_app/repository/SensorRepository.java
@@ -1,6 +1,6 @@
 package com.example.stamp_app.repository;
 
-import com.example.stamp_app.entity.Sensor;
+import com.example.stamp_app.dummyData.Sensor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
